+++ src/test/java/com/alibaba/ttl/usesjava8/ForkJoinTest.java
@@ -1,8 +1,10 @@
-package com.alibaba.ttl.threadpool.agent;
+package com.alibaba.ttl.usesjava8;
 
 import com.alibaba.ttl.TransmittableThreadLocal;
 import com.alibaba.ttl.Utils;
+import org.junit.AfterClass;
 import org.junit.Assert;
+import org.junit.Test;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -21,7 +23,7 @@ import static org.junit.Assert.assertEquals;
 /**
  * @author Jerry Lee (oldratlee at gmail dot com)
  */
-public class ForkJointAgentTest {
+public class ForkJoinTest {
     private static final ForkJoinPool pool = new ForkJoinPool();
 
     static final Map<String, Map<String, Object>> tag2copied = Collections.synchronizedMap(new LinkedHashMap<>());
@@ -36,15 +38,13 @@ public class ForkJointAgentTest {
         expandThreadPool(pool);
     }
 
+    @AfterClass
     public static void afterClass() throws Exception {
         pool.shutdown();
     }
 
-    public static void main(String[] args) throws Exception {
-        task();
-    }
-
-    public static void task() throws Exception {
+    @Test
+    public void task() throws Exception {
         final ConcurrentMap<String, TransmittableThreadLocal<String>> ttlInstances = createTestTtlValue();
 
         long[] numbers = LongStream.rangeClosed(1, 100).toArray();
@@ -92,6 +92,7 @@ class SumTask extends RecursiveTask<Long> {
 
     public final String tag;
     private final ConcurrentMap<String, TransmittableThreadLocal<String>> ttlInstances;
+    private final TransmittableThreadLocal.Capture capture;
 
     SumTask(long[] numbers, ConcurrentMap<String, TransmittableThreadLocal<String>> ttlInstances) {
         this(numbers, 0, numbers.length - 1, ttlInstances);
@@ -103,48 +104,54 @@ class SumTask extends RecursiveTask<Long> {
         this.to = to;
 
         this.ttlInstances = ttlInstances;
-        tag = ForkJointAgentTest.tagCounter.getAndIncrement() + "";
+        tag = ForkJoinTest.tagCounter.getAndIncrement() + "";
+        this.capture = TransmittableThreadLocal.capture();
     }
 
     @Override
     protected Long compute() {
-        // 1. Add new
-        String newChildKey = CHILD + tag;
-        TransmittableThreadLocal<String> child = new TransmittableThreadLocal<String>();
-        child.set(newChildKey);
+        return TransmittableThreadLocal.restoreAndRun(capture, new TransmittableThreadLocal.Action<Long, RuntimeException>() {
+            @Override
+            public Long act() {
+                // 1. Add new
+                String newChildKey = CHILD + tag;
+                TransmittableThreadLocal<String> child = new TransmittableThreadLocal<String>();
+                child.set(newChildKey);
 
-        TransmittableThreadLocal<String> old = ttlInstances.putIfAbsent(newChildKey, child);
-        if (old != null) {
-            throw new IllegalStateException("already contains key " + newChildKey);
-        }
-        ttlInstances.put(newChildKey, child);
+                TransmittableThreadLocal<String> old = ttlInstances.putIfAbsent(newChildKey, child);
+                if (old != null) {
+                    throw new IllegalStateException("already contains key " + newChildKey);
+                }
+                ttlInstances.put(newChildKey, child);
 
-        // 2. modify the parent key
-        String p = PARENT_MODIFIED_IN_CHILD + tag;
-        ttlInstances.get(PARENT_MODIFIED_IN_CHILD).set(p);
+                // 2. modify the parent key
+                String p = PARENT_MODIFIED_IN_CHILD + tag;
+                ttlInstances.get(PARENT_MODIFIED_IN_CHILD).set(p);
 
-        ForkJointAgentTest.tag2copied.put(tag, Utils.copied(ttlInstances));
+                ForkJoinTest.tag2copied.put(tag, Utils.copied(ttlInstances));
 
-        // ========================================================================
+                // ========================================================================
 
-        final int delta = to - from;
-        if (delta < 16) {
-            // compute directly
-            long total = 0;
-            for (int i = from; i <= to; i++) {
-                total += numbers[i];
+                final int delta = to - from;
+                if (delta < 16) {
+                    // compute directly
+                    long total = 0;
+                    for (int i = from; i <= to; i++) {
+                        total += numbers[i];
+                    }
+                    return total;
+                } else {
+                    // split task
+                    final int middle = from + delta / 2;
+
+                    SumTask taskLeft = new SumTask(numbers, from, middle, ttlInstances);
+                    SumTask taskRight = new SumTask(numbers, middle + 1, to, ttlInstances);
+
+                    taskLeft.fork();
+                    taskRight.fork();
+                    return taskLeft.join() + taskRight.join();
+                }
             }
-            return total;
-        } else {
-            // split task
-            final int middle = from + delta / 2;
-
-            SumTask taskLeft = new SumTask(numbers, from, middle, ttlInstances);
-            SumTask taskRight = new SumTask(numbers, middle + 1, to, ttlInstances);
-
-            taskLeft.fork();
-            taskRight.fork();
-            return taskLeft.join() + taskRight.join();
-        }
+        });
     }
 }
